Move create expense to new file
